Criar e definir Perfil de Administrador e suas funções
