+++ aplicacoes/gpa-mae/src/main/java/br/ufc/quixada/npi/gpa/controller/AuthenticationSuccessHandlerImpl.java
@@ -59,6 +59,8 @@ public class AuthenticationSuccessHandlerImpl implements
 			else if (grantedAuthority.getAuthority().equalsIgnoreCase(DISCENTE)){
 				
 				return "/aluno/selecao/listar";
+			}else if (grantedAuthority.getAuthority().equalsIgnoreCase(ADMINISTRADOR)){
+				return "/admin/listar";
 			}
 		}
