+++ aplicacoes/gpa-mae/src/main/webapp/WEB-INF/spring-security.xml
@@ -15,7 +15,7 @@
  		<intercept-url pattern="/aluno/**" access="hasRole('DISCENTE')"/>
  		
  		<intercept-url pattern="/coordenador/**" access="hasRole('COORDENADOR_ASSUNTOS_ESTUDANTIS')"/>
- 		
+ 		<intercept-url pattern="/admin/**" access="hasRole('ADMINISTRADOR_GPA')"/>
  		
  		<intercept-url pattern="/servidor/**" access="hasAnyRole('DOCENTE','STA')"/>
