+++ aplicacoes/gpa-mae/src/main/webapp/WEB-INF/views/fragments/bodyHeader.jsp
@@ -25,22 +25,24 @@
 			<div class="collapse navbar-collapse"
 				id="bs-example-navbar-collapse-1">
 				<ul class="nav navbar-nav navbar-right">
-					<sec:authorize  access="hasAnyRole('DOCENTE, STA')">
+					<sec:authorize  access="hasAnyRole('ADMINISTRADOR_GPA')">
 
 						<li role="presentation"><a
 							href="<c:url value="/servidor/listar" />">Listar Servidor <span
 								class="glyphicon glyphicon-list"></span>
 						</a></li>
 						<li role="presentation"><a
-							href="<c:url value="/selecao/listarPorServidor/${sessionScope.id}" />">Listar
-								Seleções <span class="glyphicon glyphicon-list"></span>
-						</a></li>
-						<li role="presentation"><a
 							href="<c:url value="/servidor/listar/alunos" />">Listar Alunos <span
 								class="glyphicon glyphicon-list"></span>
 						</a></li>
+					</sec:authorize>		
+					<sec:authorize access="hasAnyRole('DOCENTE', 'STA')">
+							<li role="presentation"><a
+							href="<c:url value="/selecao/listarPorServidor/${sessionScope.id}" />">Listar
+								Seleções <span class="glyphicon glyphicon-list"></span>
+						</a></li>
+					
 					</sec:authorize>
-
 					<sec:authorize access="hasAnyRole('DISCENTE')">
 						<input type="hidden" name="id" value="${sessionScope.id}" />
 						<li role="presentation"><a
